♻️ Refactor: CanvasController 위치 수정
