+++ Frontend/src/pages/StudioPage/StreamCanvas/CanvasContainer.tsx
